Moved drain_cap module to processing and changed import line in extremes.py
