+++ ruins/processing/drain_cap.py
